updated main file

+++ 3_Implementation/inc/calculator_operations.h
@@ -17,7 +17,7 @@
 * @param[in] operand2 
 * @return Result of operand1 + operand2
 */
-float add(float operand1, float operand2);
+int add(int operand1, int operand2);
 
 /**
 *  subtracts the operand1, operand2 and returns the result
@@ -25,7 +25,7 @@ float add(float operand1, float operand2);
 * @param[in] operand2 
 * @return Result of operand1 - operand2
 */
-float subtract(float operand1, float operand2);
+int subtract(int operand1, int operand2);
 
 /**
 *  multiply the operand1, operand2 and returns the result
@@ -33,7 +33,7 @@ float subtract(float operand1, float operand2);
 * @param[in] operand2 
 * @return Result of operand1 * operand2
 */
-float multiply(float operand1, float operand2);
+int multiply(int operand1, int operand2);
 
 
 /**
@@ -43,8 +43,8 @@ float multiply(float operand1, float operand2);
 * @return integer value of the operand1 / operand2
 * @note returns 0 for divide by 0 error
 */
-float divide(float operand1, float operand2);
+int divide(int operand1, int operand2);
 
-float bmi(float weight, float height);
+int bmi(int weight, int height);
 
 #endif  /* #define __CALCULATOR_OPERATIONS_H__ */
